lab::event: add the "source id" concept to events.
Now, every event::source has a unique identifier (retrieved with the static
childs' sources method 'get_id') that is used to set to every event sent
by an event::source its "source_id" member (retrieved with the event's
'get_source_id' method).
This will serve to easily discriminate event sources in listener.

The side effect to use the feature is to use the "SOURCE_MAKE_ID" in every
source's childs.

See the additional test for an exemple of how this is working.