test/mocks/glfw: add functions for keyboard testing.
In order to test keyboard events with a GLFW fashion, we add a specific events
thread started with GLFWinit that will run randomly and call windows' callbacks
at every ticks for every events in the keyboard event queue (see
glfw_mock_fake_key_action to send keyboard events to GLFW).

Additionaly, we added the 'glfwPollEvents' that will read every events of the
event queue and call window's callbacks for every of these events.